Distribute payfort service logic among other services
